Version uplift, code clean up

+++ src/main/java/com/itelg/docker/cawandu/composer/dashboard/DashboardContainerStateStatisticComposer.java
@@ -1,29 +1,23 @@
 package com.itelg.docker.cawandu.composer.dashboard;
 
-import java.util.Map.Entry;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Scope;
-import org.springframework.stereotype.Component;
-import org.zkoss.zk.ui.event.Events;
-import org.zkoss.zk.ui.select.annotation.Wire;
-import org.zkoss.zul.Listbox;
-import org.zkoss.zul.Listcell;
-import org.zkoss.zul.Listfooter;
-import org.zkoss.zul.Listitem;
-import org.zkoss.zul.Window;
-
 import com.itelg.docker.cawandu.composer.AbstractComposer;
 import com.itelg.docker.cawandu.composer.container.ContainerListComposer;
 import com.itelg.docker.cawandu.domain.container.ContainerFilter;
 import com.itelg.docker.cawandu.domain.container.ContainerState;
 import com.itelg.docker.cawandu.service.ContainerService;
 import com.itelg.zkoss.helper.i18n.Labels;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Scope;
+import org.springframework.stereotype.Component;
+import org.zkoss.zk.ui.event.Events;
+import org.zkoss.zk.ui.select.annotation.Wire;
+import org.zkoss.zul.*;
+
+import java.util.Map.Entry;
 
 @Component
 @Scope("request")
-public class DashboardContainerStateStatisticComposer extends AbstractComposer<Window>
-{
+public class DashboardContainerStateStatisticComposer extends AbstractComposer<Window> {
     private static final long serialVersionUID = -4998118686922227753L;
 
     @Autowired
@@ -36,13 +30,11 @@ public class DashboardContainerStateStatisticComposer extends AbstractComposer<W
     private Listfooter overallContainerCountListfooter;
 
     @Override
-    public void doAfterCompose(Window comp) throws Exception
-    {
+    public void doAfterCompose(Window comp) throws Exception {
         super.doAfterCompose(comp);
         int overallCount = 0;
 
-        for (Entry<ContainerState, Integer> containerStateCount : containerService.getContainerStateStats().entrySet())
-        {
+        for (Entry<ContainerState, Integer> containerStateCount : containerService.getContainerStateStats().entrySet()) {
             Listitem stateListitem = new Listitem();
             stateListitem.appendChild(new Listcell(Labels.getLabel(containerStateCount.getKey())));
             stateListitem.appendChild(new Listcell(String.valueOf(containerStateCount.getValue())));
@@ -53,7 +45,7 @@ public class DashboardContainerStateStatisticComposer extends AbstractComposer<W
                 ContainerListComposer.show(filter);
             });
             containerStateStatisticListbox.appendChild(stateListitem);
-            overallCount += containerStateCount.getValue().intValue();
+            overallCount += containerStateCount.getValue();
         }
 
         overallContainerCountListfooter.setLabel(String.valueOf(overallCount));
